Optimizations
Removed the rewrite-all option given to glsl-minifier. While this option
was a nice experiment, handling the name mapping in JS turned out to be
a major bottleneck.

Optimized Starcraft 3 models loading and rendering.
All the vertices and all the triangles now exist in one array/element
buffer respectively and are bound only once for every render.