Many changes
API:
+ unloadResource
+ getInstances
- registerModelInstanceHandler

Merged registerModelHandler and registerModelInstanceHandler into the
former.
Changed the way extending works, and added a demo OBJ handler in the
examples folder.

Scene saving/loading works again, however the rotations don't match.

unloadResource - you can now remove resources. This means that the
visibility of an object doesn't change the fact that it will be saved in
the scene.

Cleaned up some more, and a few small fixes.