Redesigning
In the middle of the process of taking out all of the handlers outside
of the model viewer.
The only handler that should be in the viewer itself (or rather in GL)
is the Texture handler, which was renamed to NativeTexture.

All the rest (Mdx, M3, BLPTexture, DDSTexture, TGATexture) can be
registered at will, or can be removed completely from the compiler.

All of this is mostly done, but there is still some cleaning to do.