Changes
Changed API function names:
loadResource -> load
unloadResource -> unload
unloadEverything -> clear

Changed the event mechanism to be more similar to that of the native
JavaScript events.
Every event has a type string, and a target, which is now an actual
viewer object (e.g. an AsyncModel).
For onprogress events, the onprogress related proprties are also
included (loaded, total, lengthComputable).

A side effect of this, is that now most of the viewer's API isn't
needed, because nearly all of the model/instance related functions just
redirect to the objects themselves.

The texture handling was changed - there is now a parent Texture object
that contains an actual texture implementation (NativeTexture,
BLPTextire, ...).
This changed the code, and texture handlers need to call the given
onload argument function, if they want to send  a "load" event.

There is a "loadend" event now for when something finishes loading,
whether after an error occurs, or it loaded, to match JS events.

The viewer's textures (used for the world modes) now send events like
every other texture - the client can choose to ignore them if it wants
to.