Big changes
The onmessage callback was removed in favor for EventTarget style event
listeners:
viewer.addEventListener(type, listener)

Possible types are: "render", "loadstart", "progress", "load", "error",
"loadend", and "abort".

The render listeners are called every frame, after the scene was
rendered.
It is useful if you want to use the internal WebGL buffer, before it is
cleared by the browser (e.g. to get pixel data).

The loading percentage functions were removed, since they were not
reliable.
Instead, there are two new functions:
loadingEnded() - returns a boolean whether there are any resources still
loading or not.
dependenciesLoaded(object) - checks if all the dependencies of an object
were loaded, and returns the ratio of loaded/total. Right now it only
actually checks for textures. Internal models and their textures will be
added later on.

In addition getWebGLContext() was added to get the internal WebGL
context, if it's needed for some reason.