Changes
Removed all the useless proxy functions in the main API, call them
directly from the objects.

The getters that returned copies or simplified string arrays (e.g.
getSequences, getTextureMap) no lnoger do that, they return the data in
the objects.
Using this data in a non-read-only way is not recommended.

Added AsyncModel[Instance].getOriginalSource, which gets the original
source that loaded the object, without modifications by mpqFile(), if
there were any.

The "render" event now properly sets the "this" and Event.target
parameters to the viewer.

Added getCameraPosition().