Removed Camera.attach()
Removed Camera.attach(), since this is functionality that should be
implemented by the client.
The other Camera functions give more flexibility to the client to
implement exactly what it wants.
Because of this, Spatial.getInverseWorldLocation() was changed to
Spatial.getWorldLocation(), which the client can use for e.g.
Camera.setTarget(), to emulate Camera.attach().