Small changes
Added Viewer.resetViewport(). Call it when the size of the canvas
changes (as long as it's not from a window resize, since the viewer
listens to resizes).

Removed Spatial.rotateQuat(), Spatial.setRotationQuat() and
Spatial.getRotationQuat().
Changed the signature of Spatial.rotate() and Spatial.setRotation() to
take two angles for the Z and X rotation axes, the same way they are
given to Camera.
Spatial.getRotation() returns the two angles.