Big changes.
Hopefully I will update the docs at some point...

There is no longer any notion of headers and packs.
There is now only Viewer.load() which loads a single model or texture,
and returns an AsyncModelInstance or an AsyncTexture.

The urls argument to the Viewer constructor no longer exists.
Instead, the second parameter of Viewer.load() is a function that
resolves paths.
It is very similar to the old urls.mpqFile() function, except one can
map anything to whatever one wants dynamically.
This removes the need for different modes for MPQ files, absolute urls,
and local paths - it is the responsibility of the path resolver to
resolve paths.

Viewer.load can now load files from memory.
The third parameter, which is optional, is used to override the file
type of the loaded resource. It can be used also for normal loads,
however it's required for loading from memory.
The fourth and last parameter is an optional boolean that defines
whether the load is from memory or not.
For example:
`Viewer.load(SomeArrayBuffer, myPathsResolver, ".mdx", true)`
This will instantly construct a new MDX handler model with the given
array buffer.
Textures can be created from any HTML valid source (Image, Canvas,
etc.), just define the file type as png.

Since objects loaded from memory have no string sources, they can't be
referenced as such.
Therefore, AsyncTextures now also have generated IDs, and texture
binding accepts IDs.
That is, if there is an AsyncTexture with id=5, then calling
gl.bindTexture(5) in an handler is valid.

Lots of code refactoring all over the place, hopefully I didn't
introduce too many bugs.