Added textures, found a browser bug?
Changed the way that the bone matrices are passed between the workers
and the main thread.
It is now passed "by reference", rather than copied.
This seems to result in far worse behavior than copying the buffers for
some reason (I believe the browsers are leaking heap memory).
Opened an issue on Chrome to see what's going on.