Changes
Started cleaning up stuff.
Main changes:
1) The viewer constructor no longer takes a path solver for the world
textures. Instead, call Viewer.setWorldTextures() which takes three
AsyncTexture objects for the ground, sky, and water.

2) The signature and behavior of Viewer.load() changed again.
The signature is now the following:
function load(src: Any|AnyArray, pathSolver: Function p, type:
String|StringArray[).
In other words, the first argument is either a single or an array of
something (depends on what you are doing), the second argument is the
path solver, and the third argument is an optional single or array of
strings that define the file types of the objects in the first argument.
If src is a single object, a single resource will be returned, if it is
an array, an array will be returned.
The type can be single also if the src is an array - in this case, all
of the objects in the array are assumed to share the type.
What does the type mean exactly? If it is defined, the src is assumed to
be from memory (e.g. an array buffer, an HTML Image object, etc.), and
is therefore sent directly to the handler for processing, with no server
fetches.
If a src array has some memory and server objects together, use the type
array such that every server object will have a null type.
E.g. load([serverPath, myArrayBuffer], pathSolver, [null, ".mdx"]).

3) Probably not very useful, but all async objects (AsyncModel,
AsyncModelInstance, AsyncTexture) now share the same event dispatching
mechanism as the viewer.
This means you can register event listeners directly to the objects
returned by Viewer.load().