Update
Made the M3 handler more consistent with itself, and changed some
implementation details.

Hopefully removed texture warnings under Firefox when rendering MDX
models.

peekCharArray now returns an array of chars, instead of a string.

Added the "loadstart" event to instances for consistency.

Added Node.resetTransformation().

Added Scene.clear().

Added ModelViewer.updateAndRender(), which does what it says.
Can be used for manual manipulation (e.g. if you set ModelViewer.paused
to true).

ModelViewer.whenAllLoaded -> ModelViewer.whenLoaded

Added ModelViewer.resourcesLoading
Added ModelViewer.whenAllLoaded(callback).

Added an early version of an automatic unit tester.
It will later be split to a more compact unit tester, and some form of a
utility class that can be used also without tests.
Tests come in the form of code that sets up a scene.
The tester can save all of the results of the tests as images (this is
not automatic...yet?)
When running the tests, the current image resulting from the viewer's
canvas, and the image saved on the server, are compared using
resemble.js