Changes
The ModelViewer object no longer has an internal infinite loop.
Use the update(), render(), and updateAndRender() however you like to
control the library.
This sort of defeats the purpose of the "render" event, but I'll keep it
for now, for consistency.

Removed ModelViewer.paused

Renamed ModelViewer.isViewerResource -> isResource

Renamed ModelViewer.getRenderdAsBlob -> toBlob

Removed ModelViewer.getRenderedAsUrl

Removed ModelViewer.getRenderedAsImage

Renamed ModelViewer.loadSingle to loadResource, and removed the old
loadResource.

Added the global function seededRandom(seed), which generates an RNG
mimicking Math.random, with your seed.

Added UnitTester.loadSync and loadResourceSync, that effectively mimick
ModelViewer.load and loadResource, but work in a synchronous way.
That is, the sources given will be loaded one by one, where each one is
given to your callback as it finishes loading.
Note that these are static methods, and require you to have an instance
of ModelViewer.

MdxSD used Date.now() as a counter that always goes up.
This made rendering non-deterministic, which bothered with the unit
tests.
Re-added a counter-per-instance.