Update
Lots of cleaning.

The MDX and M3 parsers can now throw errors, which will be reported
properly by error handlers.
This handles automatically errors when trying to load legit invalid
models.

Added the option to set a basis matrix for nodes.
This allows to change the coordinate system of nodes without affecting
instances that are attached to them.
With this change, the MDX and M3 models now reside in the same
coordinate system as the rest of the viewer.
That is, +X is right, +Y is up, and +Z goes out of the screen.

Removed Node.getLocation(), Node.getRotation(), and Node.getScaling(),
they were useless wrappers.

Fixed node attaching when attaching an instance to a node with a
non-zero pivot.
This was apparently bugged for a very, very long time, and I never
tested it in the past 2 years or so.
There is now a unit test that checks it.

Updated the MDX sanity tester.

Stumbled on a previosuly unknown sequence data behavior for MDX models.
Many models should now render more accuratly (for a specific example,
see HeroMountainKing).

Added NotifiedNode, which is a normal Node, but overrides the notify()
method to recalculate its own transformation.
It is used by instances (and other things) so that they would update
automatically when their parent updates.

Node.parent isn't stored anymore, instead there is Node.rootNode, which
is a child of the given parent.
This allows to transform this rootNode to change coordinate systems (see
above), while not changing the transformation of a parent instance.

ModelViewer.clear(deleteResources) - clear now handles both reseting the
viewer visually (removes all scenes, model views, instances, etc.), and,
if the given argument is true-like, also removing all references to all
of the actual resources.

Updated all of the unit tests, since the coordinate system changes made
all MDX and M3 tests fail.

The MDX parser now supports unsupported chunks, by giving a subreader,
much like the M3 parser.
This is useless, but whatever.