Update
Moved mix() from common.js to mix.js.
This allows to use mix() in a non-browser environment.
This was done to allow to make a Node sanity tester.
I might do the same to the rest of common.j, e.g. move the atlas
functions to their own file, etc.

Enabled simple point culling again.
This invalidated the mdx-particle-emitter test.