JassContext + many changes and fixes
Added a WIP version of a bunch of Jass related tools.
At the heart of it there is the JassContext class that is used to
effectively run Jass code in JavaScript.
Together with it is a partial implementation of the Jass natives,
variable types, and so on.
Thanks to @Ralle for sharing his Jison Jass grammar.

There are many smaller changes and fixes as well.