Big update
ES6-ified most of the code, and removed backwards compatiblity with ES5,
in favor of properly using new features such as async/await and
generators.

Added the MDLX handler, a more modern Warcraft 3 model parser, that can
read and write both the MDX and MDL formats.
It is not used by the viewer itself yet.

Many small changes, fixes, and cleaning.