Moved tests
The tests folder was always a client, so it's now in the clients folder.
This also allows it to use the same localOrHive as the other clients, which in turn allows to mostly run the unit tests on any computer for the time being. Note that it fails to fetch the relevant SC2 files, and so those tests will fail.