Lua context
The hacky JassContext and all related code was replaced with a new Lua-based context.
The Jass parser was replaced with a simple and quick function that converts Jass to Lua. Note that it assumes the Jass is valid, this isn't a syntax checker. That being said, the Lua VM will report syntax errors in the resulting Lua code.
Threads are handled correctly using Lua coroutines. This means they can support thread data such as GetExpiredTimer(), and waits using TriggerSleepAction().