Removed all texture-atlas generation
- Texture atlases are no longer needed.
They are not needed for particle emitters type 2, since all geometry emitters are now rendered per-instance.
They are not needed for the map viewer's water, because it was never needed for it.
And finally, they are not needed for the map viewer's tilesets, because the tileset textures are now used directly.
The latter is needed for reforged tileset textures, which are 4 times as big, and wouldn't fit in a texture smaller than 8k, which doesn't have great support.
On the bright side, the original mipmaps can now be used.

- Since texture atlases don't need to be generated, textures no longer need to store the useless image data object until the end of time.

- Shaders are no longer stored in the viewer. Instead, the handlers store them directly.

- Updated the version to denote it is v5 beta.