Changed the way MDX team colors/glows load for #58
The MDX handler itself now loads team colors/glows when it is added. This An optional path solver can be supplied, unless the loadsare handled in the server manually.
In addition, an optional boolean can be supplied to select between RoC/TFT (default) and Reforged teams.
All in all: viewer.addHandler(mdxHandler, pathSolver, reforgedTeams)

MdxModel no longer handles loading the team textures, nor does it matter if a specific model is a RoC/TFT or a Reforged model, only the handler determines which team colors are used.

Hopefully I didn't break any of the clients.