Parser objects no longer load from the constructor
The constructors of all of the parsers (aside from M3 which is really old) no longer accept any arguments, and no longer implicitly call the load function if arguments are given.

This allows to load corrupted files, loading whatever data can be loaded, and then throwing an exception which can be handled by the caller.

If load is called through the constructor, an exception will discard the parser entirely and no data will be salvaged.

This is required because the game supports it, and TriggerHappy reported many maps with such issues (e.g. w3i files)