Attempt to load broken MDX files
As noted a couple of commits ago, there are models that miss data and still work in the game.

The MDX handler can again attempt to load such models. If a model is missing critical information, the handler will throw an exception when it eventually fails to load it. If the handler loads the model, then the missing information wasn't important and all is fine.

I've added the same to the sanity test client, however when such a model is encountered, the loading error is reported in the test title. Whatever data was loaded will be tested normally, assuming the missing data doesn't cause the test itself to crash.